Update approach used to obtain principal ID in the API project
In the current incarnations of ASP.NET Core, B2C, etc, the user ID is being returned by default by the SUSI policy as the sub claim. ASP.NET Core is mapping this to the Name Identifier ID claim on the Claims Principal. As a result, the Home Object ID claim that the original code was looking for ("oid") is not present. The application seems to work, in that items are added/removed to/from the list, but in fact it is treating content with a "null" _currentPrincipalId. This can be seen by logging in as 1 user, creating a TODO item, then logging out and logging back in as another user - you will see the TODO item created by the first user in the second user's list.